Make design more responsive
This commit makes the design more responsive by applying a height to the feed
objects. This also implements an auto-fill for the cards section and allows the sidebar's
color to persist even though the contents do not. The sidebar and the feed can be made sticky
if required with some more implementation.